check existence of objects before deleting them Fixes: #4734
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/anatomist/anatomist-free/branches/4.2@48344 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/lib/commands/cDeleteAll.cc
@@ -88,15 +88,15 @@ DeleteAllCommand::doit()
     delete *iqw;
 
   set<AObject *> objs;
-  set<AObject *>::iterator io, eo=objs.end();
+  set<AObject *>::iterator io, eo;
   size_t cnt;
   do
   {
     cnt = 0;
     objs = theAnatomist->getObjects();
-    for( io=objs.begin(); io!=eo; ++io )
+    for( io=objs.begin(), eo=objs.end(); io!=eo; ++io )
     {
-      if( (*io)->CanBeDestroyed() )
+      if( theAnatomist->hasObject( *io ) && (*io)->CanBeDestroyed() )
         cnt += theAnatomist->destroyObject( *io );
     }
   }
